refactor: Refactor utility functions and improve equality checks in node operations, removed unused methods
