--- src/utils/is/primitives.ts
@@ -1,1 +1,0 @@
-export { isBoolean, isFunction, isNumber, isObject, isString, isUndefined } from 'radashi'
