+++ src/utils/is/three.ts
@@ -1,6 +1,6 @@
 import type { BufferGeometry, Camera, Color, ColorRepresentation, Fog, Light, Material, Mesh, Object3D, OrthographicCamera, PerspectiveCamera, Scene } from 'three'
 import { Layers } from 'three'
-import { isNumber, isString } from './primitives'
+import { isNumber, isString } from './typed'
 import { createTypeGuard } from './util'
 
 /**
