+++ src/utils/is/typed.ts
@@ -1,0 +1,1 @@
+export { isBoolean, isEqual, isFunction, isNumber, isObject, isString, isUndefined } from 'radashi'
